change beforeMessageId to beforeMessageID and fix unit test

+++ SKYKitChat/Classes/Core/Cache/SKYMessageCacheObject.h
@@ -30,7 +30,7 @@
 @property NSDate *sendDate;
 
 @property bool deleted;
-@property int seq;
+@property NSInteger seq;
 
 @property NSData *recordData;
